requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
beautifulsoup4>=4.12
openai>=1.30
pydantic>=2.6
python-dotenv>=1.0

# Brief generator (Seedance video + edge-tts + Pillow)
byteplus-python-sdk-v2[ark]
requests
edge-tts
Pillow

# Delivery pipeline (ledger-delivery/)
huggingface_hub>=0.25
